Added list_users test, renamed other list_user associated test suite to correct name
