+++ static/planttracer.js
@@ -678,5 +678,5 @@ $( document ).ready( function() {
 });
 
 if (typeof module != 'undefined'){
-module.exports = {list_movies_data, list_users_data, register_func, upload_movie_post, resend_func, check_upload_metadata, row_checkbox_clicked, computeSHA256, add_func, first_frame_url, purge_movie, upload_ready_function}
+module.exports = {list_movies_data, list_users_data, register_func, upload_movie_post, resend_func, check_upload_metadata, row_checkbox_clicked, computeSHA256, add_func, first_frame_url, purge_movie, upload_ready_function, list_users}
 }
